Fixed problem with chrome driver without headless flag, removed unused flags and rename page objects correctly
