+++ main/trello/PageObjectFactory.py
@@ -1,5 +1,5 @@
 from main.trello.gui.page_objects.login_page import LoginPage
-from main.trello.gui.page_objects.cards_page import CardsPage
+from main.trello.gui.page_objects.board_page import BoardPage
 from main.trello.gui.page_objects.home_page import HomePage
 
 
@@ -8,7 +8,7 @@ class PageObjectFactory:
     def __init__(self, driver):
         self.factory = {
             "login": LoginPage,
-            "board": CardsPage,
+            "board": BoardPage,
             "home": HomePage
         }
         self.driver = driver
